+++ lib/src/backend/storage/domain/usecase/dark_mode.dart
@@ -1,4 +1,4 @@
-import '../persistent_storage.dart';
+import '../repo/persistent_storage.dart';
 
 const darkModeKey = 'darkMode';
 const darkModeDefault = true;
