+++ lib/src/backend/storage/domain/usecase/window_size.dart
@@ -1,6 +1,6 @@
 import 'dart:ui';
 
-import '../persistent_storage.dart';
+import '../repo/persistent_storage.dart';
 
 Size? initializeWindowSizeUseCase(final PersistentStorage watch) {
   try {
